Make breakpoints usable from Node and installable from NPM.
