+++ .npmignore
@@ -1,0 +1,5 @@
+app/*
+breakpoints/*
+gh-pages/*
+tests/*
+webstore-assets/*
